feat: HyperlinkText 로직 수정
